+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * [API Change] When loading connections from a `.ini` file via `%sql --section section_name`, the section name is set as the connection alias
 * [API Change] Starting connections from a `.ini` file via `%sql [section_name]` has been deprecated
 * [Doc] Fixes documentation inaccuracy that said `:variable` was deprecated (we brought it back in `0.9.0`)
+* [Fix] Fix Twice message printing when switching to the current connection (#772)
 
 ## 0.9.1 (2023-08-10)